CI: Don't start Gradle Daemon
From the workflow output, we see that the build starts a Gradle daemon to speed up subsequent builds that never happen:

    Run ./gradlew build
      ./gradlew build
      shell: /usr/bin/bash -e {0}
      env:
        JAVA_HOME: /opt/hostedtoolcache/Java_Temurin-Hotspot_jdk/17.0.6-10/x64
        JAVA_HOME_17_X64: /opt/hostedtoolcache/Java_Temurin-Hotspot_jdk/17.0.6-10/x64

    Welcome to Gradle 7.5!
    Here are the highlights of this release:
     - Support for Java 18
     - Support for building with Groovy 4
     - Much more responsive continuous builds
     - Improved diagnostics for dependency resolution
    For more details see https://docs.gradle.org/7.5/release-notes.html
    Starting a Gradle Daemon (subsequent builds will be faster)
    ...

As recommended in https://github.com/actions/setup-java#caching-packages-dependencies
there's no need for that and we can remove some overhead by disabling it.

Also added a reference to the documentation of the caching setup (same link as above) and extracted the JDK version to an environment field.