Gradle: Integrate JaCoCo test coverage reporter (#200)
Enables the JaCoCo integration of the Android Gradle Plugin and adds a task for assembling a coverage report from the execution data collected by the unit test runner. The coverage report is based on the unit tests run with the build type debug of flavor "local".

A few new coverage-related actions are added to the CI verification workflow. First, a comment is pushed to the PR listing the coverage data of the modified lines. Then, a full coverage report of the whole project is uploaded as an artifact to the workflow.