WalletConnect: Fix parsing of "broken" schema format (#304)
The format was emitted by mistake from the a previous version of the client side library @concordium/wallet-connectors. The library has been fixed but the wallets retain support for it for backwards compatibility.

The Wallet accepted the format but passed it indiscriminantly to the crypto library without doing the appropriate conversion. This resulted in the library call failing without any error being surfaced to the user. Instead the parameters are just not presented on the approval screen. With this commit the conversion is performed right after the schema is parsed. Ideally error handling should be improved to avoid misleading the user, but as the app is being deprecated, it may not be worth the effort.